😐 修复 定时任务设置额外数据不支持 `long/int64` 类型参数问题

+++ framework/Furion.Pure/Schedule/Attributes/PeriodAttribute.cs
@@ -24,7 +24,7 @@ public class PeriodAttribute : TriggerAttribute
     /// 构造函数
     /// </summary>
     /// <param name="interval">间隔（毫秒）</param>
-    public PeriodAttribute(int interval)
+    public PeriodAttribute(long interval)
         : base(typeof(PeriodTrigger)
             , interval)
     {
